refactor plotter
